backend/requirements.txt
flask
flask-cors
scikit-learn
joblib
nltk
textblob
PyJWT
requests 